- usage of binary subjects for messages - private, protected, public and virtual transport definitions - (chained) blake2b hash values

+++ include/neuropil_data.h
@@ -46,7 +46,7 @@ extern "C" {
     struct np_data_conf {
         char key[255];
         enum np_data_type type;
-        uint32_t data_size;
+        size_t data_size;
     } NP_PACKED(1);
     // value gets appended to np_data instance in datablock
